+++ lib/src/observer/box_observer.dart
@@ -3,7 +3,7 @@ import 'package:positioned_scroll_observer/src/observer/layout_observer.dart';
 import 'package:positioned_scroll_observer/src/observer/observer_interface.dart';
 
 import 'scroll_extent.dart';
-import 'onstage_strategy.dart';
+import 'visibility_strategy.dart';
 import 'item_estimation.dart';
 import 'util.dart';
 
@@ -181,7 +181,7 @@ class MultiChildBoxObserver extends BoxScrollObserver<MultiChildRenderBox>
   @override
   void debugCheckOnstageItems({
     required ScrollExtent scrollExtent,
-    PredicatorStrategy strategy = PredicatorStrategy.tolerance,
+    VisibilityStrategy strategy = VisibilityStrategy.tolerance,
   }) {
     List<int> onstageItems = [];
 
@@ -207,9 +207,10 @@ class SingleChildBoxObserver extends BoxScrollObserver<SingleChildRenderBox>
   bool isRevealed(
     int index, {
     required ScrollExtent scrollExtent,
-    PredicatorStrategy strategy = PredicatorStrategy.tolerance,
+    VisibilityStrategy strategy = VisibilityStrategy.tolerance,
     bool shouldNormalized = true,
     bool shouldConvert = false,
+    double tolerance = 0,
   }) =>
       renderVisible;