+++ lib/src/observer/observer.dart
@@ -1,4 +1,4 @@
-export './onstage_strategy.dart';
+export 'visibility_strategy.dart';
 export './scroll_extent.dart';
 export './box_observer.dart';
 export './sliver_observer.dart';